bevy: tempory remove in game cleanup
the current implementation has a problem: it cannot run correctly on
startup, because the texture it expects as an argument, does not exist
yet. So far, bevy just gave a warning and did not run the function, but
in the browser this prevents the game for running.
Fix this later by making the texture always available but wrapped in an
Optional.